+++ R/connectors.R
@@ -10,8 +10,8 @@
 #' # Create connectors objects
 #'
 #' con <- connectors(
-#'   sdtm = connector_fs$new(path = tempdir()),
-#'   adam = connector_dbi$new(drv = RSQLite::SQLite())
+#'   sdtm = ConnectorFS$new(path = tempdir()),
+#'   adam = ConnectorDBI$new(drv = RSQLite::SQLite())
 #' )
 #'
 #' # Print for overview